Refactor nodeStatus
