+++ frontend/packages/console-app/src/status/node.ts
@@ -1,0 +1,4 @@
+import { K8sResourceKind } from '@console/internal/module/k8s';
+import { isNodeReady } from '@console/shared';
+
+export const nodeStatus = (node: K8sResourceKind) => (isNodeReady(node) ? 'Ready' : 'Not Ready');
